refactor cli emit modules
